Creating a concern for frontend logic Spina
